% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AG2020.R
\name{AG2020}
\alias{AG2020}
\title{GMPE function for Abrahamson and Gülerce et al.(2020)}
\usage{
AG2020(Mag, Rrup, Prd, ftype=0, Vs30, z25, ztor, region=7)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{ftype}{fytpe=0 for interface, ftype=1 for intraslab}

\item{Vs30}{Vs30(m/s).}

\item{z25}{.}

\item{ztor}{(km).}

\item{region}{regional flag, region=7 for Taiwan, region=8 for Global.}
}
\value{
A list will be return, including mag, evType, rRup, vs30, z25, ztor, region, mu, sigma, phi, tau, rockPGA, specT, period2, iFlag, ACadjfac, epiflag.
}
\description{
\code{AG2020} returns the ground-motion prediction with it sigma of Abrahamson and Gülerce et al.(2020) GMPE.
}
\details{
Norman Abrahamson and Zeynep Gülerce(2020) Regionalized Ground-Motion Models for
Subduction Earthquakes Based on the NGA-SUB Database, PEER Report No. 2020/25.
\url{http://dx.doi.org/10.1193/051712EQS188MR}
}
\examples{
AG2020(6, 20, 0, 0, 760, 1, 10, 7)
AG2020(7, 10, 0, 0, 760, 1, 10, 7)

}
