% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.AG2020}
\alias{cnames.AG2020}
\title{Variable name of AG2020 GMPEs.}
\format{Vectors of AG2020 GMPE.}
\usage{
cnames.AG2020
}
\description{
Variable name of AG2020 GMPEs.
}
\keyword{datasets}
