% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.KBCG2019}
\alias{cnames.KBCG2019}
\title{Variable name of KBCG2019 GMPEs.}
\format{Vectors of KBCG2019 GMPE.}
\usage{
cnames.KBCG2019
}
\description{
Variable name of KBCG2019 GMPEs.
}
\keyword{datasets}
