% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.PSHAB2019}
\alias{cnames.PSHAB2019}
\title{Variable name of PSHAB2019 GMPEs.}
\format{Vectors of PSHAB2019 GMPE.}
\usage{
cnames.PSHAB2019
}
\description{
Variable name of PSHAB2019 GMPEs.
}
\keyword{datasets}
