% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KBCG2019.R
\name{PSHAB2019}
\alias{PSHAB2019}
\title{GMPE function for Abrahamson and Gülerce et al.(2020)}
\usage{
PSHAB2019(Mag, Rrup, Prd, ftype=0, Vs30, z25, depth, Rhypo, region=7, pnwflag=1)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{ftype}{fytpe=0 for interface, ftype=1 for intraslab}

\item{Vs30}{Vs30(m/s).}

\item{z25}{.}

\item{ztor}{(km).}

\item{region}{regional flag, region=7 for Taiwan, region=8 for Global.}

\item{epiflag}{.}
}
\value{
A list will be return, including mag, Ftype, rRup, vs30, z25, lnY, sigma, phi, tau, specT, period2, iflag, depth, disthypo, iRegion, mbInter, mbSlab, pnwflag.
}
\description{
\code{PSHAB2019} returns the ground-motion prediction with it sigma of Abrahamson and Gülerce et al.(2020) GMPE.
}
\details{
Norman Abrahamson and Zeynep Gülerce(2020) Regionalized Ground-Motion Models for
Subduction Earthquakes Based on the NGA-SUB Database, PEER Report No. 2020/25.
\url{http://dx.doi.org/10.1193/051712EQS188MR}
}
\examples{
PSHAB2019(6, 20, 0, 0, 760, 1, 10, 7, 1)
PSHAB2019(7, 10, 0, 0, 760, 1, 10, 7, 1)

}
