% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI2020.R
\name{SI2020}
\alias{SI2020}
\title{GMPE function for Si et al.(2020)}
\usage{
SI2020(Mag, Rrup, Prd, ftype=0, Vs30, z25, depth, Dmoho)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{ftype}{fytpe=0 for interface, ftype=1 for intraslab}

\item{Vs30}{Vs30(m/s).}

\item{z25}{.}

\item{depth}{(km).}

\item{Dmoho}{Depth of moho(km).}

}
\value{
A list will be return, including mag, Ftype, rRup, vs30, z25, lnSa, sigma, phi, tau, specT, period2, iFlag, depth, mohodepth
}
\description{
\code{SI2020} returns the ground-motion prediction with it sigma of Si et al.(2020) GMPE.
}
\details{
Si, H., Midorikawa, S., & Kishida, T. (2022). Development of NGA-Sub ground-motion prediction equation of 5%-damped 
pseudo-spectral acceleration based on database of subduction earthquakes in Japan. Earthquake Spectra, 38(4), 2682-2706.
\url{http://dx.doi.org/10.1193/051712EQS188MR}
}
\examples{
SI2020(6, 20, 0, 0, 760, 1, 10, 7, 1)
SI2020(7, 10, 0, 0, 760, 1, 10, 7, 1)

}
